pyproject.toml
[project]
name = "werss"
version = "1.1.5"
description = "微信公众号热度分析系统"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.5.2",
    "APScheduler==3.11.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "bs4==0.0.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography==44.0.3",
    "docx2pdf==0.1.8",
    "exceptiongroup==1.3.0",
    "fastapi==0.115.12",
    "frozenlist==1.8.0",
    "greenlet==3.1.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "lxml==6.0.2",
    "Markdown==3.9",
    "markdownify==1.2.0",
    "multidict==6.7.0",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pillow==10.4.0",
    "playwright==1.55.0",
    "playwright-stealth==2.0.0",
    "propcache==0.4.1",
    "psutil==7.1.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "pyee==13.0.0",
    "PyJWT==2.8.0",
    "PyMySQL==1.1.1",
    "PySocks==1.7.1",
    "socksio>=1.0.0",
    "python-docx==1.2.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.7",
    "PyYAML==6.0.2",
    "qrcode==8.2",
    "reportlab==4.4.3",
    "requests==2.32.5",
    "schedule==1.2.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "SQLAlchemy==2.0.40",
    "starlette==0.44.0",
    "tqdm==4.67.1",
    "trio==0.27.0",
    "trio-websocket==0.12.2",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "uvicorn==0.33.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
    "yarl==1.22.0",
    "openai>=2.11.0",
    "minio>=7.2.0",
    # Visualization dependencies
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.0",
    "numba>=0.58.0",
    "pynndescent>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
# AI 提取功能（DeepSeek API，需要 openai 模块）
ai = [
    "openai>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["apis", "core", "driver", "jobs", "schemas", "tools", "qtserver"]

[tool.uv]
# uv 配置

# 配置国内镜像源以加速下载（清华镜像）
[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

# 备用：如果清华镜像失败，可以使用阿里云镜像
# [[tool.uv.index]]
# name = "aliyun"
# url = "https://mirrors.aliyun.com/pypi/simple"
